ShopProject/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
coverage==5.5
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.1.6
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.2
duplicity==0.8.12.0
Faker==6.6.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gyp==0.1
httpie==2.2.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
iniconfig==1.1.1
jeepney==0.4.3
keyring==21.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.14.0
macaroonbakery==1.3.1
Mako==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
mixer==7.1.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.7.1
pexpect==4.6.0
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.12.3
py==1.10.0
pycairo==1.16.2
pycups==2.0.1
Pygments==2.3.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
SecretStorage==3.1.2
simplejson==3.17.0
six==1.15.0
sqlparse==0.4.1
systemd-python==234
text-unidecode==1.3
toml==0.10.2
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.29+ds
wadllib==1.3.4
xkit==0.0.0
zipp==1.0.0
